fix(storefront): require a real name on the offer popup, plus in-flight fixes
Offer popup: the name input was labelled "(optional)" and the endpoint
validated it with `required: false`, while the client checked only email
and phone -- so a lead could land with no name at all, or with "John123",
"evil.com" or "<script>" in the column an admin reads.

The input is now required (minlength 2, maxlength 100) and the Alpine
component validates it before submitting, mirroring App\Rules\PersonName
message-for-message so a name the popup accepts is never bounced back by
the endpoint.

Server-side the requirement is bound to the source that renders the input
rather than to the endpoint: this one route also serves the exit-intent
popup, which has no name field, and requiring a name for every source
would have rejected every exit-intent signup. The charset guard still
applies to every source whenever a name is present, so a direct POST
cannot launder junk in by claiming a different source.

Verified by running the same 33-case table through both the PHP rule and
the JS mirror -- 0 expectation failures, 0 client/server disagreements --
and the controller rule array through Laravel's Validator across 11
source/name combinations.

Also carries the working-tree changes in flight alongside it.

+++ app/Http/Controllers/FlashSaleController.php
@@ -30,7 +30,9 @@ class FlashSaleController extends Controller
             'flashSale' => $flashSale,
             'isLive' => $hasStarted && ! $hasEnded,
             'hasEnded' => $hasEnded,
-            'remainingSeconds' => $hasEnded ? 0 : max(0, now()->diffInSeconds($flashSale->ends_at, false)),
+            // Carbon 3 returns a float here, and the raw value is printed straight
+            // into the countdown, so without the cast the seconds box reads "31.87421".
+            'remainingSeconds' => $hasEnded ? 0 : (int) max(0, now()->diffInSeconds($flashSale->ends_at, false)),
         ]);
     }
 }
